pom update

+++ src/main/java/usermanagement/ListFoodServlet.java
@@ -8,11 +8,8 @@ import jakarta.servlet.http.HttpSession;
 import org.json.JSONObject;
 import usermanagement.db.DBFoodList;
 
-
 import java.io.IOException;
 import java.io.PrintWriter;
-import java.time.LocalDate;
-import java.util.ArrayList;
 import java.util.List;
 
 @WebServlet("/listfood")
@@ -24,14 +21,13 @@ public class ListFoodServlet extends HttpServlet {
         HttpSession s = req.getSession();
         Object o = s.getAttribute("id"); // daca pe sesiune exista obiectul numit id sau nu exista voi lua diferite decizii
 
-        if(o!=null)
-        {
+        if (o != null) {
 
-            Integer i = (Integer)o;
-            int iduser = (int)i;
+            Integer i = (Integer) o;
+            int iduser = i;
             String search = req.getParameter("search");
-            if(search==null)
-                search="";
+            if (search == null)
+                search = "";
 
             DBFoodList db = new DBFoodList();
             List<MyFoodList> l = db.getFoodList(iduser, search);
@@ -39,9 +35,7 @@ public class ListFoodServlet extends HttpServlet {
             json.put("listFromBackend", l); // only food name
             String result = json.toString();
             returnJsonResponse(resp, result);
-        }
-        else
-        {
+        } else {
             error(resp, "operation forbidden. user is not logged in.");
         }
     }
@@ -59,7 +53,7 @@ public class ListFoodServlet extends HttpServlet {
         pr.close();
     }
 
-    private void error( HttpServletResponse resp, String mesaj) {
+    private void error(HttpServletResponse resp, String mesaj) {
 
         try {
             PrintWriter pw = resp.getWriter();
